separate utility functions from functions using gfe-db models
