+++ gfe-db/pipeline/config/source-config.json
@@ -1,6 +1,6 @@
 {
-    "created_utc": "2023-06-13T23:31:50Z",
-    "updated_utc": "2023-06-13T23:31:50Z",
+    "created_utc": "2023-07-09T22:55:58Z",
+    "updated_utc": "2023-07-09T22:55:58Z",
     "repositories": {
         "ANHIG/IMGTHLA": {
             "owner": "ANHIG",
