+++ gfe-db/pipeline/functions/check_source_update/constants.py
@@ -1,6 +1,4 @@
 import os
-import boto3
-from awsparameters import AppConfig
 from gfedbmodels.constants import (
     session,
     infra,
