+++ gfe-db/pipeline/lambda_layers/gfe_db_models/gfedbmodels/types.py
@@ -7,9 +7,9 @@ import jmespath
 # ExecutionState is changed using Step Functions DynamoDB states
 # SKIPPED: never processed (set by CheckSourceUpdate) ✅
 # PENDING: state machine execution started (set by CheckSourceUpdate) ✅
-# IN_PROGRESS: batch build job triggered (set by state machine)
-# SUCCESS: state machine execution succeeded (set by state machine)
-# FAILED: state machine execution failed (set by state machine)
+# IN_PROGRESS: batch build job triggered (set by state machine) ✅
+# SUCCESS: state machine execution succeeded (set by state machine) ✅
+# FAILED: state machine execution failed (set by state machine) ✅
 
 # TODO remove None, default is NOT_PROCESSED
 valid_statuses = [
@@ -174,6 +174,10 @@ class ExecutionStateItem(BaseModel):
     repository: RepositoryConfig
     commit: Commit
     execution: ExecutionDetailsConfig
+
+    @classmethod
+    def from_execution_state_item(cls, item):
+        pass
 
 
 class ExecutionState(BaseModel):
